requirements.txt
google-auth
beautifulsoup4
django-sendgrid-webhook
stripe
django>=2.2,<3.0
requests
django-spurl
sendgrid
sentry-sdk
django-suit
python-dotenv
psycopg2-binary
invoke
django-extensions
factory-boy
django-enumfield
django-anymail
django-storages[google]
gunicorn
django-credit-cards
django-filter
django-rest-knox
django-rest-passwordreset
git+https://github.com/leotrubach/django-rest-framework-proxy@fad224805c207c871cde0a394ee5eae7fb9c3750
redis>=3.3,<3.4
celery[redis]<5.0
django-celery-beat

django-maintenancemode

django-rest-framework

django-libsass
lxml

aiohttp[speedups]
asgiref

pytest
pytest-django
pytest-xdist

Collectfast>=1.3,<2.0
django-cors-headers

python-twitter
django-rest-swagger

django-redis

timezonefinder

selenium
pandas